requirements.txt
paho-mqtt
qrcode
pillow